requirements.txt
langchain==0.1.13
langchain-community==0.0.29
langchain-text-splitters==1.1.1
langchain-groq==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.13.2
pymupdf==1.24.2
streamlit==1.32.0
python-dotenv==1.0.0
groq==0.4.2
huggingface-hub==0.36.2
transformers==4.40.0
